most of the machinery is there for movement. it's not actually working yet, but something is getting to screen and blipping around.

+++ src/RoadMapNode.cpp
@@ -6,7 +6,7 @@ void RoadMapNode::addConnection(shared_ptr<RoadMapNode> conn)
 	connections.push_back(conn);
 }
 
-shared_ptr<RoadMapNode> RoadMapNode::getConnection(unsigned int idx)
+shared_ptr<RoadMapNode> RoadMapNode::getConnection(size_t idx)
 {
 	if(idx >= 0 && idx < connections.size())
 	{
@@ -17,7 +17,7 @@ shared_ptr<RoadMapNode> RoadMapNode::getConnection(unsigned int idx)
 		return nullptr;
 	}
 }
-int RoadMapNode::getNumConnections()
+size_t RoadMapNode::getNumConnections()
 {
 	return connections.size();
 }
